refactor swap form / allow transitions between the pools & the swap forms
