+++ package.json
@@ -32,6 +32,7 @@
     "next": "11.0.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-spring": "^9.2.4",
     "react-toastify": "^7.0.4",
     "recoil": "^0.4.0",
     "styled-components": "^5.3.1",
